Refactored images names and classes in Viewer
