+++ src/main/java/HollowKnight/view/elements/monsters/MinhoteMonsterViewer.java
@@ -1,7 +1,8 @@
-package HollowKnight.view.elements;
+package HollowKnight.view.elements.monsters;
 
 import HollowKnight.gui.GUI;
 import HollowKnight.model.game.elements.enemies.MinhoteMonster;
+import HollowKnight.view.elements.ElementViewer;
 import HollowKnight.view.sprites.Sprite;
 
 import java.io.IOException;
@@ -10,7 +11,7 @@ public class MinhoteMonsterViewer implements ElementViewer<MinhoteMonster> {
     private final Sprite minhoteMonsterSprite;
 
     public MinhoteMonsterViewer() throws IOException {
-        this.minhoteMonsterSprite = new Sprite("sprites/Enemies/enemy3.png");
+        this.minhoteMonsterSprite = new Sprite("sprites/Enemies/MinhoteMonster.png");
     }
 
     @Override
