+++ src/main/java/HollowKnight/view/elements/monsters/PurpleMonsterViewer.java
@@ -1,7 +1,8 @@
-package HollowKnight.view.elements;
+package HollowKnight.view.elements.monsters;
 
 import HollowKnight.gui.GUI;
 import HollowKnight.model.game.elements.enemies.PurpleMonster;
+import HollowKnight.view.elements.ElementViewer;
 import HollowKnight.view.sprites.Sprite;
 
 import java.io.IOException;
@@ -10,7 +11,7 @@ public class PurpleMonsterViewer implements ElementViewer<PurpleMonster> {
     private final Sprite purpleMonsterSprite;
 
     public PurpleMonsterViewer() throws IOException {
-        this.purpleMonsterSprite = new Sprite("sprites/Enemies/Enemy2.png");
+        this.purpleMonsterSprite = new Sprite("sprites/Enemies/PurpleMonster.png");
     }
 
     @Override
