+++ src/main/java/HollowKnight/view/elements/rocks/BigRockViewer.java
@@ -1,7 +1,8 @@
-package HollowKnight.view.elements;
+package HollowKnight.view.elements.rocks;
 
 import HollowKnight.gui.GUI;
 import HollowKnight.model.game.elements.rocks.BigRock;
+import HollowKnight.view.elements.ElementViewer;
 import HollowKnight.view.sprites.Sprite;
 
 import java.io.IOException;
@@ -10,7 +11,7 @@ public class BigRockViewer implements ElementViewer<BigRock> {
     private final Sprite bigRockSprite;
 
     public BigRockViewer() throws IOException {
-        bigRockSprite = new Sprite("sprites/Ambient/bigrock.png");
+        bigRockSprite = new Sprite("sprites/Ambient/Bigrock.png");
     }
     @Override
     public void draw(BigRock model, GUI gui) throws IOException {
