+++ src/main/java/HollowKnight/view/elements/trees/SmallTreeViewer.java
@@ -1,7 +1,8 @@
-package HollowKnight.view.elements;
+package HollowKnight.view.elements.trees;
 
 import HollowKnight.gui.GUI;
 import HollowKnight.model.game.elements.Tree.SmallTree;
+import HollowKnight.view.elements.ElementViewer;
 import HollowKnight.view.sprites.Sprite;
 
 import java.io.IOException;
@@ -10,7 +11,7 @@ public class SmallTreeViewer implements ElementViewer<SmallTree> {
     private final Sprite TreeSprite;
 
     public SmallTreeViewer() throws IOException {
-        TreeSprite = new Sprite("sprites/Ambient/tree2.png");
+        TreeSprite = new Sprite("sprites/Ambient/SmallTree.png");
     }
     @Override
     public void draw(SmallTree model, GUI gui) throws IOException {
