+++ src/main/resources/sprites/Ambient/MediumTree.png
