+++ src/main/resources/sprites/Ambient/SmallTree.png
