+++ src/main/resources/sprites/Ambient/Smallrock.png
