+++ src/main/resources/sprites/Enemies/MinhoteMonster-reversed.png
