+++ src/main/resources/sprites/Enemies/MinhoteMonster.png
