+++ src/main/resources/sprites/Enemies/PurpleMonster.png
